+++ pyrightconfig.json
@@ -1,4 +1,6 @@
 {
-    "exclude": ["src/flask_ml/flask_ml_server/_generated_models.py"]
-  }
-  
+  "python.analysis.typeCheckingMode": "basic",
+  "exclude": [
+    "src/flask_ml/flask_ml_server/_generated_models.py"
+  ]
+}
